domain: build 1.2.14 specific stuff only if libvirt.1.2.14 tag is present
