+++ README.md
@@ -5,9 +5,21 @@ Go bindings for libvirt.
 Make sure to have `libvirt-dev` package (or the development files otherwise somewhere in your include path)
 
 ## Version Support
-Currently, the only supported version of libvirt is **1.2.2**, tagged as `v2.x` releases `gopkg.in/alexzorin/libvirt-go.v2` [(docs)](http://gopkg.in/alexzorin/libvirt-go.v2).
 
-The bindings will probably work with versions of libvirt that are higher than 1.2.2, depending on what is added in those releases. However, no features are currently being added that will cause the build or tests to break against 1.2.2.
+The minimum supported version of libvirt is **1.2.2**. Due to the
+API/ABI compatibility promise of libvirt, more recent versions of
+libvirt should work too.
+
+Some features require a more recent version of libvirt. They are
+disabled by default. If you want to enable them, build using one of
+those additional tags (you need to use only the most recent one you
+are interested in):
+
+ - **1.2.14**
+
+For example:
+
+    go build -tags libvirt.1.2.14
 
 ### OS Compatibility Matrix
 
@@ -16,9 +28,11 @@ To quickly see what version of libvirt your OS can easily support (may be outdat
 | OS Release   | libvirt Version                |
 | ------------ | ------------------------------ |
 | FC19         | 1.2.9 from libvirt.org/sources |
-| Debian 7     | 1.2.4 from wheezy-backports    |
-| Debian 6     | 0.9.12 from squeeze-backports  |
+| Debian 8     | 1.2.9 from jessie              |
+| Debian 7     | 1.2.9 from wheezy-backports    |
 | Ubuntu 14.04 | 1.2.2 from trusty              |
+| Ubuntu 16.04 | 1.3.1 from xenial              |
+| RHEL 7       | 1.2.17                         |
 | RHEL 6       | 0.10.x                         |
 | RHEL 5       | 0.8.x                          |
